Better control pointer alignments.
It seems that either llvm-hs or LLVM itself assumes that a pointer
load or store that doesn't indicate an explicit alignment is aligned
to the size of pointee.  But this is wrong for vector loads from
multidimensional Dex arrays---the minor dimension need only be aligned
on the size of a single array element, not the whole vector.