Teach occurrence analysis to hoist using the free variables it already computed, instead of re-traversing the term.
Pull out a little common machinery between this and DCE, where we also
do this optimization.

On the BFGS example, this speeds up occurrence analysis by 60-70%, and
(according to the profiler) is worth something like a 7% improvement
to compilation time end-to-end.

Also make two minor improvements to classes relating to WriterT1 that
came up while experimenting with this.