Detect method types that are disallowed for automatic instance derivation.
The current implementation of automatic instance derivation synthesizes
methods (for the derived instance) in a way that is guided by the method
types in the instance that we are deriving _from_. For the method synthesis
to work correctly certain types (i.e. the type we are deriving _from_ and
the type we are deriving an instance _for_) must not be mentioned in the
type of the class method that corresponds to the instance method that is
being synthesized.