Put peephole optimizations in one place and use them from Builder.
The reason to do this now is that I want to make AD linearity explicit for
correctness reasons. I started doing that but realized I was going to need
linear version of each of the helper functions in Builder `add`, `mul`. This
cuts down on that boilerplate and it's a good idea anyway.