Rewrite the live view system to be explicitly incremental by dealing in diffs.
Previously we achieved incremental updates using a sort of "rebuild the world
but memoize all the things" strategy. I've come around to the idea that it's
better to deal with updates explicitly.

The new pattern could be called "defunctionalized state". You maintain a state
but instead of being able to update it arbitrarily (e.g. using `put` in the
state monad) you can only update it in a certain number of predefined ways,
expressed as a "diff" data type. These diffs can be sent between servers,
serialized to JSON and sent to the browser and so on. And they're expected to be
monoidal so you can batch a bunch of diffs together for efficiency and even hope
for some cancellation, say of pushes and pops.

The other part of the pattern is to build "state servers" that serve the current
state using these diffs. When you subscribe to them they send you a current
state and an endless stream of updates. You can wire the servers together to
achieve an incrementally updated pipeline.

This worked out really well! I was even able to add a new feature: tracking the
current state of a cell (pending/running/complete) in the web view. And the
whole thing is much snappier because every piece of it is just working with
updates instead of full states.