Replace source spans with IDs.
These are better because we can hang more information off them. We can give an
ID to e.g. the binder of a `Ix n` constraint that doesn't actually have a source
span and then choose how to highlight later on depending on the situation. And
we can use the IDs to model the graph of relationships between source
components.